+++ tests/testthat/test-7-ssh_proxy.r
@@ -112,9 +112,13 @@ test_that("full SSH connection", {
     sched = getOption("clustermq.scheduler", qsys_default)
     skip_if(is.null(sched) || toupper(sched) != "MULTICORE",
             message="options(clustermq.scheduler') must be 'MULTICORE'")
-
     options(clustermq.template = "SSH", clustermq.ssh.host="127.0.0.1")
+
     w = workers(n_jobs=1, qsys_id="ssh", reuse=FALSE)
     result = Q(identity, 42, n_jobs=1, timeout=10L, workers=w)
     expect_equal(result, list(42))
+
+    w = workers(n_jobs=2, qsys_id="ssh", reuse=FALSE)
+    result = clustermq::Q(Sys.sleep, time=c(1,2), n_jobs=2)
+    expect_equal(result, list(NULL, NULL))
 })