+++ NEWS.md
@@ -1,5 +1,6 @@
 # git head
 
+* Fix a bug where SSH proxy would not cache data properly (#320)
 * Local parallelism (`multicore`, `multiprocess`) again uses local IP (#321)
 
 # clustermq 0.9.1